+++ src/tools/sysup/commands/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import annotations
-
-from .mac.mac import CommandMac
-from .nvim.nvim import CommandNvim
-from .pip.pip import CommandPip
-from .step_result import StepResult
-from .wsl.wsl import CommandWsl
-
-__all__ = ["CommandMac", "CommandNvim", "CommandPip", "CommandWsl", "StepResult"]