+++ src/tools/sysup/commands/nvim/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from .nvim import CommandNvim
-
-__all__ = ["CommandNvim"]