+++ src/tools/sysup/commands/pip/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-from .pip import CommandPip
-
-__all__ = ["CommandPip"]